Major Changes
Changed alot of things.

Namely, got rid of the lock object and just put all roboclaw calls in the same thread, called when the topic receives a value.

Also went through and removed things that were for the drivetrain and changed things to match the arm.

Also added inverting direction and soft limit functionality to the pid controller.