Added armStates to the node
Before the plan was to expect another node to constantly supply the setpoint to this node. This could have unexpected results so now this node can receive an input from the topic and it has a loop to constantly recalculate the PID values.

Also added 2 ways to disable the arm. One will safely move the arm to the home position and disable. The other will abruptly stop all motors and can be used in emergency cases.